Move matching to its own module
